1) create a dir for swarm. 2) Use Ev instead of Result<Option<Ev>> for function  3) Some document for swarm and swarm builder
